fixing merge conflicts

+++ server/routes/taskApi.js
@@ -3,6 +3,16 @@ const { v4: uuidv4 } = require("uuid"); // creates unique ids for the todos
 const router = express.Router();
 const pool = require("../db");
 
+// create middleware
+const authMiddleware = (req, res, next) => {
+  const token = req.headers["x-token"];
+  if (!token) return res.status(401).json({ detail: "Token not present" });
+  const { email } = jwt.decode(token, process.env.JWT_SECRET);
+  if (!email) return res.status(401).json({ detail: "Unauthorized" });
+  req.email = email;
+  next();
+};
+
 /**
  * @swagger
  * tags:
@@ -28,7 +38,7 @@ const pool = require("../db");
  *         date:
  *           type: string
  *         completed:
- *           type: boolean 
+ *           type: boolean
  */
 
 // Get all todos
@@ -57,13 +67,11 @@ const pool = require("../db");
  *                 $ref: '#/components/schemas/Todo'
  */
 
-router.get("/:userEmail", async (req, res) => {
-  const { userEmail } = req.params;
-
+router.get("/", authMiddleware, async (req, res) => {
   try {
     const todos = await pool.query(
       "SELECT * FROM todos WHERE user_email = $1",
-      [userEmail]
+      [req.email]
     );
     res.json(todos.rows);
   } catch (err) {
@@ -94,18 +102,25 @@ router.get("/:userEmail", async (req, res) => {
  *               $ref: '#/components/schemas/Todo'
  */
 
-router.post("/", async (req, res) => {
+router.post("/", authMiddleware, async (req, res) => {
   try {
-    const { user_email, title, progress, date, completed } = req.body;
-    console.log(user_email, title, progress, date, completed);
+    const { title, progress, date, completed } = req.body;
     const id = uuidv4();
     const newTodo = await pool.query(
       `INSERT INTO todos(id, user_email, title, progress, date, completed) VALUES($1,$2,$3,$4,$5,$6)`,
-      [id, user_email, title, progress, date, completed]
+      [
+        id,
+        [id, req.email, title, progress, date],
+        title,
+        progress,
+        date,
+        completed,
+      ]
     );
     res.json(newTodo);
   } catch (err) {
     console.error(err.message);
+    res.status(500).send({ err });
   }
 });
 
@@ -139,17 +154,26 @@ router.post("/", async (req, res) => {
  *               $ref: '#/components/schemas/Todo'
  */
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", authMiddleware, async (req, res) => {
   const { id } = req.params;
-  const { user_email, title, progress, date, completed } = req.body;
   try {
+    const todo = await pool.query("SELECT * FROM todos WHERE id=$1", [id]);
+    if (!todo) return res.status(404).json({ detail: "Not found" });
+    if (todo.rows[0].user_email !== req.email) {
+      return res
+        .status(401)
+        .json({ detail: "You're not authorized to modify this resource" });
+    }
+
+    const { title, progress, date } = req.body;
     const editTodo = await pool.query(
       "UPDATE todos SET user_email = $1, title = $2, progress = $3, date = $4, completed = $5 WHERE id = $6;",
-      [user_email, title, progress, date, completed, id]
+      [req.email, title, progress, date, id]
     );
     res.json(editTodo);
   } catch (err) {
     console.error(err.message);
+    res.status(500).send({ err });
   }
 });
 
@@ -177,15 +201,26 @@ router.put("/:id", async (req, res) => {
  *               type: object
  */
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", authMiddleware, async (req, res) => {
   const { id } = req.params;
   try {
-    const deleteTodo = await pool.query("DELETE FROM todos WHERE id = $1;", [
+    const todo = await pool.query("SELECT * FROM todos WHERE id = $1", [id]);
+    if (!todo) return res.status(404).json({ detail: "Not found" });
+
+    if (todo.rows[0].user_email !== req.email) {
+      return res
+        .status(401)
+        .json({ detail: "You're not authorized to delete this resource" });
+    }
+
+    const deleteToDo = await pool.query("DELETE FROM todos WHERE id = $1", [
       id,
     ]);
-    res.json(deleteTodo);
+
+    res.json(deleteToDo);
   } catch (err) {
     console.error(err.message);
+    res.status(500).send({ err });
   }
 });
